Finish Gibbs Sampling

+++ src/main/java/MRLDA.java
@@ -11,6 +11,9 @@ import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
 import org.apache.hadoop.mapreduce.lib.output.MultipleOutputs;
 import org.apache.hadoop.mapreduce.lib.output.TextOutputFormat;
 import utils.DocumentInputFormat;
+import utils.Multinomial;
+
+import static utils.Settings.*;
 
 import java.io.IOException;
 
@@ -21,39 +24,52 @@ public class MRLDA {
 
     public static void main(String args[]) throws Exception {
         Configuration config = new Configuration();
-
-        //INIT JOB
-        Job job_init = Job.getInstance(config, "MR_LDA_Init");
-        job_init.setJarByClass(MRLDA.class);
-        job_init.setMapperClass(InitMapper.class);
-        job_init.setMapOutputKeyClass(Text.class);
-        job_init.setMapOutputValueClass(Text.class);
-        //job_init.setCombinerClass(UnifiedReducer.class);
-        job_init.setReducerClass(UnifiedReducer.class);
-        MultipleInputs.addInputPath(job_init, new Path("/users/rocks5/13307130228/corpus/corpus.cp")
-                , DocumentInputFormat.class);
-        MultipleOutputs.addNamedOutput(job_init,"WC", TextOutputFormat.class, Text.class, Text.class);
-        MultipleOutputs.addNamedOutput(job_init,"WZ", TextOutputFormat.class, Text.class, Text.class);
-        FileOutputFormat.setOutputPath(job_init, new Path("/users/rocks5/13307130228/result"));
         FileSystem fileSystem = FileSystem.get(config);
-        fileSystem.delete(new Path("/users/rocks5/13307130228/result"), true);
-        job_init.waitForCompletion(true);
 
-        //FIRST ITERATION JOB
-        Job job_iteration = Job.getInstance(config, "MR_LDA_Iteration");
-        job_iteration.setJarByClass(MRLDA.class);
-        job_iteration.setMapperClass(GSMapper.class);
-        job_iteration.setMapOutputKeyClass(Text.class);
-        job_iteration.setMapOutputValueClass(Text.class);
-        job_iteration.setReducerClass(UnifiedReducer.class);
-        MultipleInputs.addInputPath(job_iteration, new Path("/users/rocks5/13307130228/temp/WZ/")
-                , DocumentInputFormat.class);
-        MultipleOutputs.addNamedOutput(job_iteration,"WC", TextOutputFormat.class, Text.class, Text.class);
-        MultipleOutputs.addNamedOutput(job_iteration,"WZ", TextOutputFormat.class, Text.class, Text.class);
-        FileOutputFormat.setOutputPath(job_iteration, new Path("/users/rocks5/13307130228/result"));
-        fileSystem.delete(new Path("/users/rocks5/13307130228/result"), true);
+        if (args[1].equals("restart")) {
+            //Clear /temp/C and /temp/WZ
+            fileSystem.delete(new Path("/users/rocks5/13307130228/temp/C/"), true);
+            fileSystem.delete(new Path("/users/rocks5/13307130228/temp/WZ/"), true);
 
-        System.exit(job_iteration.waitForCompletion(true) ? 0 : 1);
+            //INIT JOB
+            Job job_init = Job.getInstance(config, "MR_LDA_Init");
+            job_init.setJarByClass(MRLDA.class);
+            job_init.setMapperClass(InitMapper.class);
+            job_init.setMapOutputKeyClass(Text.class);
+            job_init.setMapOutputValueClass(Text.class);
+            job_init.setReducerClass(UnifiedReducer.class);
+            MultipleInputs.addInputPath(job_init, new Path("/users/rocks5/13307130228/corpus/corpus.cp")
+                    , DocumentInputFormat.class);
+            MultipleOutputs.addNamedOutput(job_init, "WC", TextOutputFormat.class, Text.class, Text.class);
+            MultipleOutputs.addNamedOutput(job_init, "WZ", TextOutputFormat.class, Text.class, Text.class);
+            FileOutputFormat.setOutputPath(job_init, new Path("/users/rocks5/13307130228/result"));
+            fileSystem.delete(new Path("/users/rocks5/13307130228/result"), true);
+            job_init.waitForCompletion(true);
+        }
+        int IterationCount = Iteration_No;
+        if (Iteration_No == -1) {
+            IterationCount = Integer.parseInt(args[0]);
+        }
+        for (int i = 0; i < IterationCount; i++) {
+            //Gibbs Sampling ITERATION JOB
+            Job job_iteration = Job.getInstance(config, "MR_LDA_Iteration");
+            job_iteration.setJarByClass(MRLDA.class);
+            job_iteration.setMapperClass(GSMapper.class);
+            job_iteration.setMapOutputKeyClass(Text.class);
+            job_iteration.setMapOutputValueClass(Text.class);
+            job_iteration.setReducerClass(UnifiedReducer.class);
+            MultipleInputs.addInputPath(job_iteration, new Path("/users/rocks5/13307130228/temp/WZ/")
+                    , DocumentInputFormat.class);
+            MultipleOutputs.addNamedOutput(job_iteration, "WC", TextOutputFormat.class, Text.class, Text.class);
+            MultipleOutputs.addNamedOutput(job_iteration, "WZ", TextOutputFormat.class, Text.class, Text.class);
+            FileOutputFormat.setOutputPath(job_iteration, new Path("/users/rocks5/13307130228/result"));
+            fileSystem.delete(new Path("/users/rocks5/13307130228/result"), true);
+            if (i != IterationCount - 1)
+                job_iteration.waitForCompletion(true);
+            else
+                System.exit(job_iteration.waitForCompletion(true) ? 0 : 1);
+        }
+
     }
 
 }
